Fixed #29052 -- Create test databases before migrate
Use a list() to store the test database aliases so the order remains
mostly stable by following the order of the connections. Always use
the default database alias as the first alias to accommodate migrate.

Previously migrate could be executed on a secondary alias on a separate
connection/transaction which caused data migrations to fail because they
would write to the default database.

Introduce a new function ``initialize_test_db`` which can be executed
after all test databases have been created and all settings patched and
it is safe for migrations to access any database they want.